[utils] moved functions operating on dicts to utlis
