+++ src/giskardpy/config_loader.py
@@ -4,7 +4,7 @@ from copy import deepcopy
 import rospy
 import yaml
 
-from giskardpy.utils import get_ros_pkg_path
+import giskardpy.utils as utils
 
 
 class Loader(yaml.SafeLoader):
@@ -19,7 +19,7 @@ class Loader(yaml.SafeLoader):
         try:
             self.config_root = os.path.split(stream.name)[0]
             self.ros_package_keywords = [u'ros://', u'package://']
-            self.giskardpy_root = get_ros_pkg_path(u'giskardpy')
+            self.giskardpy_root = utils.get_ros_pkg_path(u'giskardpy')
         except AttributeError:
             self.config_root = os.path.curdir
 
@@ -36,7 +36,7 @@ def get_filename(loader, node, root):
             raise SyntaxError(u'Invalid ros package path: please use ros:// or package:// as path prefix.')
         removed_key_word = file_or_ros_path_str.replace(loader.ros_package_keywords[indices[0]], '')
         path_split = removed_key_word.split('/')
-        package_path = get_ros_pkg_path(path_split[0])
+        package_path = utils.get_ros_pkg_path(path_split[0])
         filename = package_path + removed_key_word.replace(path_split[0], '')
     else:
         filename = os.path.abspath(os.path.join(root, file_or_ros_path_str))
@@ -61,95 +61,17 @@ def construct_find(loader, node):
     return get_filename(loader, node, loader.giskardpy_root)
 
 
-def update_nested_dicts(d, u):
-    """
-    Will update the values in the nested dict d from nested dict u and 
-    add new key-value-pairs from nested dict u into nested dict d." \
-
-    :type d: dict
-    :type u: dict
-    :rtype: dict
-    :returns: updated nested dict
-    """
-    for k, v in u.items():
-        if type(v) == dict and d.get(k, {}) is not None:
-            d[k] = update_nested_dicts(d.get(k, {}), v)
-        else:
-            d[k] = v
-    return d
-
-
-def find_parent_of_key(key, value, keys):
-    """
-    Will return the values of the dict managing the given key and
-    returns the key-chain leading from the nested dict value to the
-    dict containing the given key.
-
-    :type key: str
-    :type value: dict
-    :type keys: list(str)
-    :rtype: tuple(dict, list(str))
-    :returns: tuple containing a dict and the key-chain in a list
-    """
-    if key in value:
-        yield value, keys
-    else:
-        for k, v in value.items():
-            if type(v) == dict:
-                keys.append(k)
-                yield find_parent_of_key(key, v, keys)
-
-
-def nested_update(dic, keys, value):
-    """
-    Will update the nested dict dic with the key-chain keys with the given value.
-
-    :type dic: dict
-    :type keys: list(str)
-    :type value: dict or str
-    :rtype: dict
-    :returns: updated dict
-    """
-    if keys:
-        for key in keys[:-1]:
-            dic = dic.setdefault(key, {})
-        dic[keys[-1]].update(value)
-    else:
-        dic.update(value)
-
-
-def update_parents(d, merge_key='parent'):
-    """
-    Will merge the dict containing the given key merge_key with the value in d[merge_key].
-
-    :type d: dict
-    :returns: dict
-    """
-    root_data = deepcopy(d)
-    gen = find_parent_of_key(merge_key, root_data, [])
-    while True:
-        try:
-            data, keys = next(gen)
-        except (StopIteration, ValueError):
-            break
-        parent_data = data[merge_key]
-        data.pop(merge_key)
-        updated = update_nested_dicts(parent_data, data)
-        nested_update(root_data, keys, updated)
-    return root_data
-
-
 def load_robot_yaml(path):
     with open(path, 'r') as f:
         data = yaml.load(f, Loader)
-        return update_parents(data)
+        return utils.update_parents(data)
 
 
 def ros_load_robot_config(config_file, test=False):
-    config = load_robot_yaml(get_ros_pkg_path(u'giskardpy') + u'/config/' + config_file)
+    config = load_robot_yaml(utils.get_ros_pkg_path(u'giskardpy') + u'/config/' + config_file)
     if test:
-        config = update_nested_dicts(deepcopy(config),
-                                     load_robot_yaml(get_ros_pkg_path(u'giskardpy') + u'/config/test.yaml'))
+        config = utils.update_nested_dicts(deepcopy(config),
+                                     load_robot_yaml(utils.get_ros_pkg_path(u'giskardpy') + u'/config/test.yaml'))
     if config and not rospy.is_shutdown():
         rospy.set_param('~', config)
         return True
